requirements.txt
aiogram>=3.28.1
apscheduler>=3.11.2
aiohttp>=3.13.5
python-dotenv>=1.2.2
asyncpg>=0.31.0
sqlalchemy>=2.0.49